substituindo uuidv4 por uuid

+++ src/services/DeleteProviderService.ts
@@ -1,4 +1,7 @@
 import { getRepository } from 'typeorm'
+import { validate } from 'uuid'
+
+import AppError from '@errors/AppError'
 
 import Provider from '@entity/Provider'
 
@@ -8,6 +11,8 @@ interface Request {
 
 export default class DeleteProviderService {
   async execute({ id }: Request) {
+    if (!validate(id)) throw new AppError('O ID solicitado não foi encontrado!', 404)
+
     const providerRepository = getRepository(Provider)
     await providerRepository.delete({ id })
   }
